requirements.txt
# ETH_Seed_Scanner dependencies

# Core dependencies
web3>=6.0.0
eth-account>=0.8.0
bip32utils>=0.3.post4
mnemonic>=0.20.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
click>=8.1.3
openpyxl>=3.0.10

# Security and configuration
cryptography>=41.0.0
pyyaml>=6.0.0

# Development and testing
pytest>=7.3.1
black>=23.3.0
pylint>=2.17.0
